Move ValidationException to other package
