+++ src/main/java/bisq/core/dao/exceptions/ValidationException.java
@@ -1,4 +1,4 @@
-package bisq.core.dao.governance;
+package bisq.core.dao.exceptions;
 
 import bisq.core.dao.state.blockchain.Tx;
