+++ src/main/java/bisq/core/dao/governance/proposal/confiscatebond/ConfiscateBondProposalService.java
@@ -19,7 +19,7 @@ package bisq.core.dao.governance.proposal.confiscatebond;
 
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
-import bisq.core.dao.governance.ValidationException;
+import bisq.core.dao.exceptions.ValidationException;
 import bisq.core.dao.governance.proposal.BaseProposalService;
 import bisq.core.dao.governance.proposal.ProposalConsensus;
 import bisq.core.dao.governance.proposal.ProposalWithTransaction;
